Dashboard enhancements: Color fix for PH meter Now presents meters for all measurement types for all active batches, previous limit was 2 meters Measurement edit fix

+++ src/main/java/com/brewery/dto/Gauge.java
@@ -83,7 +83,7 @@ public class Gauge {
 		return stops;
 	}
 
-	public void addStop( long value, String color ) {
+	public void addStop( double value, String color ) {
 		// Stop stop = new Stop( value, color );
 		ArrayList<Object> list = new ArrayList<>();
 		list.add( value );
